split accept model and views
