+++ Sources/UI/PaymentMethods/SepaAcceptViewController.swift
@@ -5,115 +5,11 @@
 //  Created by Uwe Tilemann on 23.11.22.
 //
 
-import Foundation
+import Combine
 import UIKit
 import SwiftUI
-import Combine
 import SnabbleCore
 
-public final class SepaAcceptModel: ObservableObject {
-    /// subscribe to this Publisher to  process
-    public var actionPublisher = PassthroughSubject<[String: Any]?, Never>()
-    
-    let process: CheckoutProcess
-    
-    init(process: CheckoutProcess) {
-        self.process = process
-    }
-    
-    public var markup: String? {
-        guard let markup = process.paymentPreauthInformation?.markup,
-              let body = markup.replacingOccurrences(of: "+", with: " ").removingPercentEncoding else {
-            return nil
-        }
-        
-        let head = """
-<html>
-        <head>
-            <meta charset="utf-8" />
-            <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no" />
-            <style type="text/css">
-                pre { font-family: -apple-system, sans-serif; font-size: 15px; white-space: pre-wrap; }
-                body { padding: 8px 8px }
-                * { font-family: -apple-system, sans-serif; font-size: 15px; word-wrap: break-word }
-                *, a { color: #000 }
-                h1 { font-size: 22px }
-                h2 { font-size: 17px }
-                h4 { font-weight: normal; color: #3c3c43; opacity: 0.6 }
-                @media (prefers-color-scheme: dark) {
-                    a, h4, * { color: #fff }
-                }
-            </style>
-        </head>
-        <body>
-"""
-        let trail  = """
-</body></html>
-"""
-        
-        return head + body + trail
-    }
-}
-
-extension SepaAcceptModel {
-    public func accept(completion: @escaping (_ result: RawResult<CheckoutProcess, SnabbleError>) -> Void ) async throws {
-                
-        print("will authorize")
-        
-//        if self.isValid,
-//           let cert = Snabble.shared.certificates.first,
-//           let sepaData = PayoneSepaData(cert.data, iban: self.iban, lastName: self.lastname, city: self.city, countryCode: self.countryCode) {
-//
-//            let detail = PaymentMethodDetail(sepaData)
-//            PaymentMethodDetails.save(detail)
-//
-//            paymentDetail = detail
-//
-//            DispatchQueue.main.async {
-//                self.objectWillChange.send()
-//            }
-//        } else {
-//            throw PaymentMethodError.encryptionError
-//        }
-    }
-}
-
-public struct SepaAcceptView: View {
-    @ObservedObject public var model: SepaAcceptModel
-
-    public init(model: SepaAcceptModel) {
-        self.model = model
-    }
-    
-    @ViewBuilder
-    var text: some View {
-        if let markup = model.markup {
-            HTMLView(string: markup)
-        } else {
-            Text(keyed: "Snabble.SEPA.mandate")
-        }
-    }
-    
-    @ViewBuilder
-    var button: some View {
-        Button(action: {
-            model.actionPublisher.send(["action": "accept"])
-        }) {
-            Text(keyed: "Snabble.SEPA.iAgree")
-                .frame(maxWidth: .infinity)
-        }
-        .buttonStyle(AccentButtonStyle())
-    }
-    
-    public var body: some View {
-        VStack {
-            text
-            button
-        }
-        .padding()
-    }
-    
-}
 /// Methods for managing callbacks for widges
 public protocol SepaAcceptViewControllerDelegate: AnyObject {
 
@@ -160,9 +56,7 @@ extension SepaAcceptViewController: SepaAcceptViewControllerDelegate {
     func accept(model: SepaAcceptModel) {
         Task {
             do {
-                if let handler = completionHandler {
-                    try await model.accept(completion: handler)
-                }
+                try await model.accept(completion: completionHandler ?? { _ in })
 
                 DispatchQueue.main.async {
                     self.dismiss()
@@ -171,7 +65,9 @@ extension SepaAcceptViewController: SepaAcceptViewControllerDelegate {
                 DispatchQueue.main.async {
                     let alert = AlertView(title: nil, message: Asset.localizedString(forKey: "Snabble.SEPA.encryptionError"))
                     
-                    alert.alertController?.addAction(UIAlertAction(title: Asset.localizedString(forKey: "ok"), style: .default))
+                    alert.alertController?.addAction(UIAlertAction(title: Asset.localizedString(forKey: "ok"), style: .default) { _ in
+                        self.dismiss()
+                    })
                     alert.show()
                 }
             }
